Documented and cleaned up code

+++ 2D_Game_Engine/io/console.cpp
@@ -1,3 +1,12 @@
+///////////////////////////////////////////////////////////////
+// io/console.cpp
+// 
+// This is the console class, which handles everything needed for the console. It 
+// handles input, selecting and storing characters, displaying the right textures, 
+// and generally displaying graphics as appropriate. It is static, though needs to 
+// be initialized. 
+///////////////////////////////////////////////////////////////
+
 #include "console.h"
 #include <SDL_keycode.h>
 #include "fonts.h"
@@ -11,20 +20,37 @@ static SDL_Color textColor, backgroundColor, cursorColor;
 static std::string textFont;
 static int backgroundOpacity;
 static int numLines, maxLines;
+//Width and height of our cursor
 static int cursorWidth, cursorHeight;
 static bool showConsole;
+//String containing the current command's text
 static std::string currentCommand;
+//Object for displaying the console background
 static Object* backgroundObj;
+//List of objects which make up the text on our console
 static std::list<Object*>* textObjMap;
 static Object* cursor;
+//Is the user holding shift
 static bool shift;
+//The amount to move the background down upon every command run
 static int backgroundShiftIncrementSize;
+//The Y location of the background
 static int requestedBackgroundY;
 
 Console::Console() {}
 
 Console::~Console() {}
 
+/**
+ * Initializes the console
+ * 
+ * @param int text_h					Text height
+ * @param int text_w					Text width
+ * @param SDL_Color text_color			Text color
+ * @param std::string text_font			Text font (must be a supported font)
+ * @param SDL_Color background_color	Console background color
+ * @param int background_opacity		Console background opacity
+ */
 void Console::init(int text_h, int text_w, SDL_Color text_color, std::string text_font, 
 				   SDL_Color background_color, int background_opacity) {
 	textHeight = text_h;
@@ -39,22 +65,32 @@ void Console::init(int text_h, int text_w, SDL_Color text_color, std::string tex
 	cursorHeight = 16;
 	showConsole = false;
 	backgroundShiftIncrementSize = (textHeight + 1);
+
+	//Set up the background
 	backgroundObj = new Object();
 	backgroundObj->setTexture(Texture::getTexture("assets/textures/console_background.png"));
 	backgroundObj->setDestination(0, -((textHeight + 1) * maxLines), Game::getWindowWidth(), ((textHeight + 1) * maxLines));
+	
 	textObjMap = new std::list<Object*>();
+
 	currentCommand = "";
+
+	//Set up the cursor
 	cursor = new Object();
 	cursor->setX(1);
 	cursor->setY(1);
 	cursor->setWidth(cursorWidth);
 	cursor->setHeight(cursorHeight);
+
 	//Set the cursor color equal to the text color, for now
 	cursorColor = text_color;
 	shift = false;
 	requestedBackgroundY = -((textHeight + 1) * maxLines);
 }
 
+/**
+ * Opens the console
+ */
 void Console::openConsole() {
 	//Keyboard inputs will automatically go to the console
 
@@ -70,25 +106,38 @@ void Console::openConsole() {
 	showConsole = true;
 }
 
+/**
+ * Closes the console
+ */
 void Console::closeConsole() {
 	//Stop showing console
 	showConsole = false;
 	backgroundObj->setDestination(-2, -2, -1, -1);
 }
 
-//TODO: Testing
+/**
+ * Writes a character to the console. 
+ * 
+ * @param int sym, ASCII key code
+ */
 void Console::writeChar(int sym) {
+
+	//Apply shift if it is currently held
 	if (shift) {
+		//TODO: This only works on letters at the moment. Expand functionality
+		//to include more keys. 
 		sym -= 32;
 	}
 
+	//Add this character to our string
 	currentCommand += (char)sym;
 
 	std::string cString{(char)sym};
 
 	//Create our character object
 	Object* newChar = new Object();
-	newChar->setDestination((textWidth+1) * (currentCommand.length()-1), (numLines-1) * (textHeight+1), textWidth, textHeight);
+	newChar->setDestination((textWidth+1) * (currentCommand.length()-1), (numLines-1) * 
+							(textHeight+1), textWidth, textHeight);
 	
 	//Get the character texture
 	newChar->setTexture(Fonts::getRenderedText(cString, textFont, textHeight, textColor));
@@ -102,6 +151,9 @@ void Console::writeChar(int sym) {
 	cursorRight();
 }
 
+/**
+ * Creates a new line in the console. This should occur when enter/return is pressed. 
+ */
 void Console::newLine() {
 	//Run the command present in the text bar
 	Commands::runCommand(currentCommand);
@@ -214,16 +266,22 @@ SDL_Color Console::getCursorColor() {
 	return cursorColor;
 }
 
+/**
+ * Backspace function. Gets called when the user presses backspace!
+ */
 void Console::backspace() {
 
+	//Move the cursor left once
 	cursorLeft();
+
 
 	if (currentCommand.length() > 0) {
 		//Remove the character to the left of the cursor
-		int charToRemove = ((cursor->getX() - 1) / (textWidth + 1));
+		int charToRemove = (((int)cursor->getX() - 1) / (textWidth + 1));
 		currentCommand.erase(charToRemove, 1);
 		/*
-		//Get the iterator for the beginning of our command, and find the element just before our cursor
+		//Get the iterator for the beginning of our command, and find the element 
+		//just before our cursor
 		std::list<Object*>::iterator it = textObjMap->begin();
 		for (int i = 0; i < charToRemove; i++) {
 			it++;
@@ -272,11 +330,18 @@ void Console::backspace() {
 	}
 }
 
+/**
+ * Sets whether shift is on or not. True if it's on, false if it's off. 
+ */
 void Console::setShift(bool b) {
 	shift = b;
 }
 
+/**
+ * Updates the console. 
+ */
 void Console::update() {
+	//Update the position of the background object. 
 	if (backgroundObj->getDestination()->y < requestedBackgroundY) {
 		backgroundObj->update();
 	}
